Moved HomeScreen logic into its own component.
